move test out of misc now
