+++ test/test_misc.py
@@ -1,6 +1,5 @@
 import diffrax
 import jax.numpy as jnp
-import pytest
 
 from .helpers import tree_allclose
 
@@ -10,13 +9,3 @@ def test_fill_forward():
     out_ = jnp.array([jnp.nan, 0.0, 1.0, 1.0, 1.0, 2.0, 2.0])
     fill_in = diffrax._misc.fill_forward(in_[:, None])
     assert tree_allclose(fill_in, out_[:, None], equal_nan=True)
-
-
-def test_weaklydiagonal_deprecate():
-    with pytest.warns(
-        PendingDeprecationWarning,
-        match="WeaklyDiagonalControlTerm is pending deprecation",
-    ):
-        _ = diffrax.WeaklyDiagonalControlTerm(
-            lambda t, y, args: 0.0, lambda t0, t1: jnp.array(t1 - t0)
-        )
